be/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
python-jose
python-multipart
passlib[bcrypt]
PyJWT
pydantic[email]
python-dotenv